cell and row mods become private. Row and Cell are reexported in crate's root
