+++ examples/basic.rs
@@ -1,8 +1,6 @@
 #[macro_use]
 extern crate prettytable;
-use prettytable::Table;
-use prettytable::row::Row;
-use prettytable::cell::Cell;
+use prettytable::{Table, Row, Cell};
 
 /*
     Following main function will print :
