+++ examples/span.rs
@@ -1,6 +1,6 @@
 #[macro_use]
 extern crate prettytable;
-use prettytable::{row::Row, cell::Cell, format::Alignment};
+use prettytable::{Row, Cell, format::Alignment};
 
 
 fn main() {
