+++ src/main.rs
@@ -2,8 +2,8 @@
 #[macro_use]
 extern crate prettytable;
 use prettytable::Table;
-use prettytable::row::Row;
-use prettytable::cell::Cell;
+use prettytable::Row;
+use prettytable::Cell;
 use prettytable::format::*;
 use prettytable::{Attr, color};
